Refactor room subscription logic
Make it so that the only users who are subscribed to a room get the
message that was sent. Also revert the Public Room implementation
as there was no sense in having it, since supabase does not support
filtering channels by a null row. Might implement it back when
supabase supports it.